requirements.txt
google-play-scraper==1.2.4
app-store-scraper==0.3.5
sqlalchemy==2.0.25
pydantic==2.5.3
python-dotenv==1.0.0
pandas==2.1.4
pytest==7.4.4
streamlit==1.29.0
openai==1.6.1
beautifulsoup4==4.12.2
